feat: 금액 관련 타입 BigDecimal로 변경, 통합테스트, 레포지토리 테스트용 부모 클래스 생

+++ src/main/java/kr/hhplus/be/server/application/order/OrderFacade.java
@@ -23,6 +23,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Component
@@ -47,7 +48,7 @@ public class OrderFacade {
                 .map(OrderProductsRequest::toDto)
                 .toList();
 
-        int totalPrice = productService.getTotalPriceBy(orderProductsDtos);
+        BigDecimal totalPrice = productService.getTotalPriceBy(orderProductsDtos);
         productService.decreaseProductStock(orderProductsDtos);
 
         Long userCouponId = orderRequest.userCouponId();
