+++ src/main/java/kr/hhplus/be/server/domain/order/dto/OrderResult.java
@@ -1,12 +1,14 @@
 package kr.hhplus.be.server.domain.order.dto;
 
+import java.math.BigDecimal;
+
 public record OrderResult(
         Long orderId,
         Long userId,
         Long couponId,
         Long userCouponId,
-        int basePrice,
-        int discountAmount,
-        int finalPrice
+        BigDecimal basePrice,
+        BigDecimal discountAmount,
+        BigDecimal finalPrice
 ) {
 }
