+++ src/main/java/kr/hhplus/be/server/interfaces/api/order/response/OrderResponse.java
@@ -2,13 +2,15 @@ package kr.hhplus.be.server.interfaces.api.order.response;
 
 import kr.hhplus.be.server.domain.order.dto.OrderResult;
 
+import java.math.BigDecimal;
+
 public record OrderResponse(
         Long id,
         Long userId,
         Long couponId,
-        int basePrice,
-        int discountAmount,
-        int finalPrice
+        BigDecimal basePrice,
+        BigDecimal discountAmount,
+        BigDecimal finalPrice
 ) {
 
     public static OrderResponse from(OrderResult orderResult) {
