+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/response/PointChargeResponse.java
@@ -2,9 +2,11 @@ package kr.hhplus.be.server.interfaces.api.point.response;
 
 import kr.hhplus.be.server.domain.point.dto.PointChargeResult;
 
+import java.math.BigDecimal;
+
 public record PointChargeResponse(
         Long userId,
-        int amount
+        BigDecimal amount
 ) {
 
     public static PointChargeResponse from(PointChargeResult pointChargeResult) {
