+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/response/PointSearchResponse.java
@@ -2,9 +2,11 @@ package kr.hhplus.be.server.interfaces.api.point.response;
 
 import kr.hhplus.be.server.domain.point.dto.PointSearchResult;
 
+import java.math.BigDecimal;
+
 public record PointSearchResponse(
         Long userId,
-        int balance
+        BigDecimal balance
 ) {
 
     public static PointSearchResponse from(PointSearchResult searchResult) {
